move model assembling to lca
